DRY out `hasOwnProperty` by extracting a lower-level module.
