+++ local-modules/typecheck/package.json
@@ -4,6 +4,8 @@
   "main": "main.js",
 
   "dependencies": {
-    "object-inspect": "^1.2.1"
+    "object-inspect": "^1.2.1",
+
+    "util-base": "local"
   }
 }
