+++ local-modules/util-base/README.md
@@ -1,0 +1,15 @@
+util-base
+=========
+
+Super-base utility functionality. This module only exists so as to avoid
+a circular dependency between the modules `typecheck` and `util-common`.
+In general, it shouldn't be used directly. Its functionality is more "publicly"
+exposed via `util-common`.
+
+- - - - - - - - - -
+
+```
+Copyright 2016-2017 the Bayou Authors (Dan Bornstein et alia).
+Licensed AS IS and WITHOUT WARRANTY under the Apache License,
+Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
+```
