+++ local-modules/util-base/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "util-base",
+  "version": "1.0.0",
+  "main": "main.js"
+}
